Code refactoring WIP

+++ src/ESC.h
@@ -19,13 +19,13 @@ enum ESCId{ESC0, ESC1, ESC2, ESC3};
 class ESC {
  public:
     // Power setup
-    const uint8_t MIN_POWER = 1060;
+    const int MIN_POWER = 1060;
     // Max power available to stabilize quardirotors - Set to 1860 to reach max power
-    const uint8_t MAX_POWER = 1860;
-    const uint8_t MAX_THROTTLE_PERCENT = 100.0;
+    const int MAX_POWER = 1860;
+    const int MAX_THROTTLE_PERCENT = 100.0;
     // Threshold on max throttle command (percent of max motor power)
     uint8_t MAX_THROTTLE = MAX_POWER*(MAX_THROTTLE_PERCENT/100.0);
-    uint8_t IDLE_THRESHOLD = 1100;
+    int IDLE_THRESHOLD = 1100;
 
  public:
     typeESC ESCList[4];
